+++ backend/Cargo.toml
@@ -17,7 +17,6 @@ username = []
 serde.workspace = true
 serde_json.workspace = true
 http.workspace = true
-bytes.workspace = true
 diesel.workspace = true
 argon2.workspace = true
 chrono.workspace = true
