+++ backend/src/lib.rs
@@ -1,3 +1,8 @@
+#![allow(clippy::derive_partial_eq_without_eq, clippy::get_first)]
+#![deny(unused_crate_dependencies)]
+#![warn(clippy::used_underscore_binding)]
+#![forbid(unsafe_code)]
+
 extern crate self as email_auth;
 
 mod handlers;
